requirements.txt
seaborn
scipy
matplotlib
omegaconf
fastcluster
torch
torchvision
numpy
pandas
hydra-core
sklearn
tqdm
